backend/requirements.txt
fastapi[standard]
uvicorn[standard]
python-dotenv
ibm-watsonx-ai
pydantic